Fix unbound keys during reload
Sometimes Phoenix hasn't freed the previous key handlers during reload,
resulting in the keys not being bound. This applies a timeout to try and rebind
any unbound keys during the intial loading of the config.

A timeout must be used, because if we occupy the main thread the keys will never
be freed.