Use osascript for internal display brightness
The `brightness` command-line utility (nriley/brightness) does not work
correctly. It doesn't set the actual system brightness, which results
in the display eventually reverting the brightness setting. It also does
not match the value set via the system brightness slider.

This is a hacky, somewhat tolerable solution that launches System
Preferences (Displays) in the background and sets the brightness via the
slider. To reduce delay, System Preferences is launched as soon as the
brightness adjustment has begun, and quit after the setting has been
applied.